refactor(components): move reusable components to a separate folder :recycle:
