+++ src/index.php
@@ -20,8 +20,8 @@
 
 <body>
 <div data-role="page">
-  <?php require './sidebar.php'?><!-- /panel -->
-  <?php require './header.php'?><!-- header -->
+  <?php require './components/sidebar.php'?><!-- /panel -->
+  <?php require './components/header.php'?><!-- header -->
   <div role="main" class="overlay ui-content main-content">
     <div class="ui-grid-a ui-responsive">
       <div class="ui-block-a">
@@ -85,7 +85,7 @@
       </div>
     </div><!-- /grid-a -->
   </div><!-- /content -->
-  <?php require './footer.php'?><!--footer -->
+  <?php require './components/footer.php'?><!--footer -->
 </div><!-- page -->
 <!-- inject:js -->
 <!-- endinject -->
